grafana: relax uid requirement
So far grizzly enforce that the name strictly equals to the uid and even
set the UID in the parse function.

This is problematic because some characters are allowed in the name
and not in the uid. For instance many mixins put a ".json" has a suffix
in the name and the "." character is not something allowed in an UID.

With this commit grizzly will now only enforce that the uid exist and is
not an empty string. It also set by default to the md5 of the name in
the parse function.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>